restructure vue codebase
