+++ client/packages/codebase/src/components/Project/Step/Configuration.vue
