+++ client/packages/codebase/src/components/Project/Step/MetadataEditor.vue
