+++ client/packages/codebase/src/components/Project/Step/Publish.vue
