+++ client/packages/codebase/src/components/Settings/ThemeSelector.vue
